requirements.txt
Pillow==6.2.1
numpy==1.17.3
opencv-python==4.1.1.26